requirements.txt
Django>=3.0.7
djangorestframework>=3.11.0
sentry-sdk>=0.14.4
gunicorn>=20.0.4
Pillow>=7.1.2
django-cors-headers>=3.3.0
psycopg2>=2.8.5
cloudinary>=1.21.0
python-dotenv>=0.13.0
django-better-admin-arrayfield>=1.1.0
django-filter>=2.3.0
django-admin-menu>=1.3.0
twilio>=6.44.0
python-memcached>=1.59
